refactor: move search components closer to search
