+++ web/src/pages/Search/components/SearchObjectEntry.tsx
