+++ web/src/pages/Search/components/styled.tsx
@@ -1,0 +1,24 @@
+import styled from 'styled-components';
+
+export const SearchInput = styled.input`
+width: 60vw;
+max-width: 640px;
+`;
+
+export const SearchContainer = styled.div`
+position: sticky;
+margin: 0 auto;
+display: flex;
+justify-content: center;
+`;
+
+export const EmptyContainer = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 50vh;
+`;
+
+export const StyledSearchPage = styled.div`
+  margin: 0 auto;
+`;
